feat: - add logging - add exceptions

+++ src/main/java/ru/itis/fisd/semestrovka/repository/PurchaseRepository.java
@@ -4,12 +4,11 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.itis.fisd.semestrovka.entity.Purchase;
-import ru.itis.fisd.semestrovka.entity.User;
-
-import java.util.List;
+import ru.itis.fisd.semestrovka.entity.orm.Purchase;
+import ru.itis.fisd.semestrovka.entity.orm.User;
 
 @Repository
-public interface PurchaseRepository extends JpaRepository<Purchase, Integer> {
+public interface PurchaseRepository extends JpaRepository<Purchase, Long> {
     Page<Purchase> findAllByUser(User user, Pageable pageable);
+
 }
